+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- Option to pcbnew_do export:
+  - `--mirror/-M`      Mirror printing (KiCad 6 only!)
+
 
 ## [1.5.1] - 2020-10-08
 ### Added